+++ app/src/js/actions/index.js
@@ -182,13 +182,13 @@ export const getGranule = (granuleId, params) => ({
   }
 });
 
-export const getGranuleRecoveryStatus = (granuleId) => ({
+export const getGranuleRecoveryStatus = (granuleId, collectionId) => ({
   [CALL_API]: {
     type: types.RECOVERY_GRANULE,
     method: 'POST',
     id: granuleId,
     path: 'orca/recovery/granules',
-    data: { granuleId }
+    data: { granuleId, collectionId }
   }
 });
 
@@ -758,7 +758,10 @@ export const updateRule = (payload) => ({
     type: types.UPDATE_RULE,
     method: 'PUT',
     path: `rules/${payload.name}`,
-    data: payload
+    data: payload,
+    headers: {
+      'Cumulus-API-Version': '2',
+    }
   }
 });
 
@@ -795,6 +798,9 @@ export const enableRule = (payload) => {
       data: {
         ...rule,
         state: 'ENABLED'
+      },
+      headers: {
+        'Cumulus-API-Version': '2',
       }
     }
   };
@@ -812,6 +818,9 @@ export const disableRule = (payload) => {
       data: {
         ...rule,
         state: 'DISABLED'
+      },
+      headers: {
+        'Cumulus-API-Version': '2',
       }
     }
   };
@@ -826,6 +835,9 @@ export const rerunRule = (payload) => ({
     data: {
       ...payload,
       action: 'rerun'
+    },
+    headers: {
+      'Cumulus-API-Version': '2',
     }
   }
 });